chore: 스프린트 2 작업 내용을 배포 브랜치로 병합한다.

+++ backend/src/main/java/com/woowacourse/naepyeon/controller/TeamController.java
@@ -1,13 +1,21 @@
 package com.woowacourse.naepyeon.controller;
 
+import com.woowacourse.naepyeon.controller.auth.AuthenticationPrincipal;
 import com.woowacourse.naepyeon.controller.dto.CreateResponse;
+import com.woowacourse.naepyeon.controller.dto.JoinTeamMemberRequest;
+import com.woowacourse.naepyeon.controller.dto.LoginMemberRequest;
 import com.woowacourse.naepyeon.controller.dto.TeamRequest;
+import com.woowacourse.naepyeon.exception.UncertificationTeamMemberException;
 import com.woowacourse.naepyeon.service.TeamService;
+import com.woowacourse.naepyeon.service.dto.JoinedMembersResponseDto;
+import com.woowacourse.naepyeon.service.dto.TeamResponseDto;
+import com.woowacourse.naepyeon.service.dto.TeamsResponseDto;
 import java.net.URI;
 import javax.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
@@ -22,22 +30,70 @@ public class TeamController {
 
     private final TeamService teamService;
 
+    @GetMapping("/{teamId}")
+    public ResponseEntity<TeamResponseDto> getTeam(
+            @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
+            @PathVariable final Long teamId) {
+        return ResponseEntity.ok(teamService.findById(teamId, loginMemberRequest.getId()));
+    }
+
+    @GetMapping("/me")
+    public ResponseEntity<TeamsResponseDto> getJoinedTeams(
+            @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest) {
+        return ResponseEntity.ok(teamService.findByJoinedMemberId(loginMemberRequest.getId()));
+    }
+
+    @GetMapping
+    public ResponseEntity<TeamsResponseDto> getAllTeams(
+            @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest) {
+        return ResponseEntity.ok(teamService.findAll(loginMemberRequest.getId()));
+    }
+
+    @GetMapping("/{teamId}/members")
+    public ResponseEntity<JoinedMembersResponseDto> getJoinedMembers(
+            @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
+            @PathVariable final Long teamId) {
+        if (!teamService.isJoinedMember(loginMemberRequest.getId(), teamId)) {
+            throw new UncertificationTeamMemberException(teamId, loginMemberRequest.getId());
+        }
+        return ResponseEntity.ok(teamService.findJoinedMembers(teamId));
+    }
+
+
     @PostMapping
-    public ResponseEntity<CreateResponse> createMember(@RequestBody @Valid final TeamRequest teamRequest) {
-        final Long teamId = teamService.save(teamRequest.getName());
+    public ResponseEntity<CreateResponse> createTeam(
+            @AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
+            @RequestBody @Valid final TeamRequest teamRequest) {
+        final Long teamId = teamService.save(teamRequest, loginMemberRequest.getId());
         return ResponseEntity.created(URI.create("/api/v1/teams/" + teamId)).body(new CreateResponse(teamId));
     }
 
     @PutMapping("/{teamId}")
-    public ResponseEntity<Void> updateMember(@PathVariable final Long teamId,
-                                             @RequestBody @Valid final TeamRequest teamRequest) {
+    public ResponseEntity<Void> updateTeam(@AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
+                                           @PathVariable final Long teamId,
+                                           @RequestBody @Valid final TeamRequest teamRequest) {
+        if (!teamService.isJoinedMember(loginMemberRequest.getId(), teamId)) {
+            throw new UncertificationTeamMemberException(teamId, loginMemberRequest.getId());
+        }
         teamService.updateName(teamId, teamRequest.getName());
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{teamId}")
-    public ResponseEntity<Void> deleteMember(@PathVariable final Long teamId) {
+    public ResponseEntity<Void> deleteTeam(@AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
+                                           @PathVariable final Long teamId) {
+        if (!teamService.isJoinedMember(loginMemberRequest.getId(), teamId)) {
+            throw new UncertificationTeamMemberException(teamId, loginMemberRequest.getId());
+        }
         teamService.delete(teamId);
+        return ResponseEntity.noContent().build();
+    }
+
+    @PostMapping("/{teamId}")
+    public ResponseEntity<Void> joinMember(@AuthenticationPrincipal @Valid final LoginMemberRequest loginMemberRequest,
+                                           @PathVariable final Long teamId,
+                                           @RequestBody final JoinTeamMemberRequest joinTeamMemberRequest) {
+        teamService.joinMember(teamId, loginMemberRequest.getId(), joinTeamMemberRequest.getNickname());
         return ResponseEntity.noContent().build();
     }
 }
