+++ backend/src/main/java/com/woowacourse/naepyeon/service/MessageService.java
@@ -3,9 +3,15 @@ package com.woowacourse.naepyeon.service;
 import com.woowacourse.naepyeon.domain.Member;
 import com.woowacourse.naepyeon.domain.Message;
 import com.woowacourse.naepyeon.domain.Rollingpaper;
+import com.woowacourse.naepyeon.domain.Team;
+import com.woowacourse.naepyeon.exception.NotAuthorException;
+import com.woowacourse.naepyeon.exception.NotFoundMemberException;
+import com.woowacourse.naepyeon.exception.NotFoundMessageException;
+import com.woowacourse.naepyeon.exception.NotFoundRollingpaperException;
+import com.woowacourse.naepyeon.repository.MemberRepository;
 import com.woowacourse.naepyeon.repository.MessageRepository;
 import com.woowacourse.naepyeon.repository.RollingpaperRepository;
-import com.woowacourse.naepyeon.repository.jpa.MemberJpaDao;
+import com.woowacourse.naepyeon.repository.TeamParticipationRepository;
 import com.woowacourse.naepyeon.service.dto.MessageResponseDto;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -20,19 +26,20 @@ public class MessageService {
 
     private final MessageRepository messageRepository;
     private final RollingpaperRepository rollingpaperRepository;
-    private final MemberJpaDao memberJpaDao; //todo: #19 머지 후 memberRepository로 변경
+    private final MemberRepository memberRepository;
+    private final TeamParticipationRepository teamParticipationRepository;
 
     public Long saveMessage(final String content, final Long authorId, final Long rollingpaperId) {
         final Rollingpaper rollingpaper = rollingpaperRepository.findById(rollingpaperId)
-                .orElseThrow(IllegalArgumentException::new);
-        final Member author = memberJpaDao.findById(authorId)
-                .orElseThrow(IllegalArgumentException::new);
+                .orElseThrow(() -> new NotFoundRollingpaperException(rollingpaperId));
+        final Member author = memberRepository.findById(authorId)
+                .orElseThrow(() -> new NotFoundMemberException(authorId));
         final Message message = new Message(content, author, rollingpaper);
         return messageRepository.save(message);
     }
 
     @Transactional(readOnly = true)
-    public List<MessageResponseDto> findMessages(final Long rollingpaperId) {
+    public List<MessageResponseDto> findMessages(final Long rollingpaperId, final Long teamId) {
         return messageRepository.findAllByRollingpaperId(rollingpaperId)
                 .stream()
                 .map(message -> {
@@ -40,28 +47,47 @@ public class MessageService {
                     return new MessageResponseDto(
                             message.getId(),
                             message.getContent(),
-                            author.getUsername(),
+                            findMessageWriterNickname(teamId, message),
                             author.getId()
                     );
                 })
                 .collect(Collectors.toUnmodifiableList());
     }
 
-    @Transactional(readOnly = true)
-    public MessageResponseDto findMessage(final Long messageId) {
-        final Message message = messageRepository.findById(messageId)
-                .orElseThrow(IllegalArgumentException::new);
-
+    private String findMessageWriterNickname(final Long teamId, final Message message) {
         final Member author = message.getAuthor();
-
-        return new MessageResponseDto(messageId, message.getContent(), author.getUsername(), author.getId());
+        return teamParticipationRepository.findNicknameByMemberId(author.getId(), teamId);
     }
 
-    public void updateContent(final Long messageId, final String newContent) {
+    @Transactional(readOnly = true)
+    public MessageResponseDto findMessage(final Long messageId, final Long rollingpaperId) {
+        final Message message = messageRepository.findById(messageId)
+                .orElseThrow(() -> new NotFoundMessageException(messageId));
+        final Rollingpaper rollingpaper = rollingpaperRepository.findById(rollingpaperId)
+                .orElseThrow(() -> new NotFoundRollingpaperException(rollingpaperId));
+        final Team team = rollingpaper.getTeam();
+        final Member author = message.getAuthor();
+        final String nickname = findMessageWriterNickname(team.getId(), message);
+        return new MessageResponseDto(messageId, message.getContent(), nickname, author.getId());
+    }
+
+    public void updateContent(final Long messageId, final String newContent, final Long memberId) {
+        final Message message = messageRepository.findById(messageId)
+                .orElseThrow(() -> new NotFoundMessageException(messageId));
+        validateAuthor(memberId, message);
         messageRepository.update(messageId, newContent);
     }
 
-    public void deleteMessage(final Long messageId) {
+    public void deleteMessage(final Long messageId, final Long memberId) {
+        final Message message = messageRepository.findById(messageId)
+                .orElseThrow(() -> new NotFoundMessageException(messageId));
+        validateAuthor(memberId, message);
         messageRepository.delete(messageId);
+    }
+
+    private void validateAuthor(final Long memberId, final Message message) {
+        if (!message.isAuthor(memberId)) {
+            throw new NotAuthorException(memberId);
+        }
     }
 }
